fix(kanban): address review followups (#186)
